Move eventing code
