+++ eventing/api/manager.go
