+++ eventing/factory.go
@@ -2,12 +2,11 @@ package eventing
 
 import (
 	"fmt"
+	"github.com/project-kessel/inventory-api/eventing/api"
+	"github.com/project-kessel/inventory-api/eventing/kafka"
+	"github.com/project-kessel/inventory-api/eventing/stdout"
 
 	"github.com/go-kratos/kratos/v2/log"
-
-	"github.com/project-kessel/inventory-api/internal/eventing/api"
-	"github.com/project-kessel/inventory-api/internal/eventing/kafka"
-	"github.com/project-kessel/inventory-api/internal/eventing/stdout"
 )
 
 func New(c CompletedConfig, source string, logger *log.Helper) (api.Manager, error) {
