+++ eventing/kafka/manager.go
@@ -8,12 +8,12 @@ import (
 	"github.com/confluentinc/confluent-kafka-go/v2/kafka"
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/google/uuid"
+	api "github.com/project-kessel/inventory-api/eventing/api"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
-	"github.com/project-kessel/inventory-api/internal/eventing/api"
 )
 
 type KafkaManager struct {
