+++ internal/biz/common.go
@@ -2,21 +2,21 @@ package biz
 
 import (
 	"context"
+	"github.com/project-kessel/inventory-api/eventing/api"
 	authzapi "github.com/project-kessel/inventory-api/internal/authz/api"
 	"github.com/project-kessel/inventory-api/internal/biz/model"
-	eventingapi "github.com/project-kessel/inventory-api/internal/eventing/api"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	"time"
 )
 
-func DefaultResourceSendEvent(ctx context.Context, model *model.Resource, eventer eventingapi.Manager, reportedTime time.Time, operationType eventingapi.OperationType) error {
+func DefaultResourceSendEvent(ctx context.Context, model *model.Resource, eventer api.Manager, reportedTime time.Time, operationType api.OperationType) error {
 	identity, err := middleware.GetIdentity(ctx)
 	if err != nil {
 		return err
 	}
 
 	producer, _ := eventer.Lookup(identity, model.ResourceType, model.ID)
-	evt, err := eventingapi.NewResourceEvent(operationType, model, reportedTime)
+	evt, err := api.NewResourceEvent(operationType, model, reportedTime)
 	if err != nil {
 		return err
 	}
@@ -29,14 +29,14 @@ func DefaultResourceSendEvent(ctx context.Context, model *model.Resource, evente
 	return nil
 }
 
-func DefaultRelationshipSendEvent(ctx context.Context, m *model.Relationship, eventer eventingapi.Manager, reportedTime time.Time, operationType eventingapi.OperationType) error {
+func DefaultRelationshipSendEvent(ctx context.Context, m *model.Relationship, eventer api.Manager, reportedTime time.Time, operationType api.OperationType) error {
 	identity, err := middleware.GetIdentity(ctx)
 	if err != nil {
 		return err
 	}
 
 	producer, _ := eventer.Lookup(identity, m.RelationshipType, m.ID)
-	evt, err := eventingapi.NewRelationshipEvent(operationType, m, reportedTime)
+	evt, err := api.NewRelationshipEvent(operationType, m, reportedTime)
 	if err != nil {
 		return err
 	}
